Move function into THREE utilities
